2-execute_a_command.pp

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -3,4 +3,5 @@
 exec { 'killing a process using pkill':
   command  => 'pkill -9 killmenow',
   path     => '/usr/bin:/bin',
+  onlyif   => 'pgrep killmenow',
 }
